# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DeployRequest": {
        "properties": {
          "strategy_id": {
            "format": "uuid",
            "title": "Strategy Id",
            "type": "string"
          }
        },
        "required": [
          "strategy_id"
        ],
        "title": "DeployRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaperSessionOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initial_cash": {
            "title": "Initial Cash",
            "type": "string"
          },
          "last_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "session_start": {
            "format": "date-time",
            "title": "Session Start",
            "type": "string"
          },
          "snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "strategy_id": {
            "format": "uuid",
            "title": "Strategy Id",
            "type": "string"
          },
          "strategy_version": {
            "title": "Strategy Version",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "id",
          "strategy_id",
          "strategy_version",
          "symbol",
          "timeframe",
          "status",
          "initial_cash",
          "session_start",
          "last_run_at",
          "snapshot"
        ],
        "title": "PaperSessionOut",
        "type": "object"
      },
...
      "PositionSizeOut": {
        "description": "Sizing GUIDANCE \u2014 never an executable order (live trading stays gated).\nAll figures are Decimal strings computed server-side from the engine's math.",
        "properties": {
          "notional": {
            "title": "Notional",
            "type": "string"
          },
          "pct_of_equity": {
            "title": "Pct Of Equity",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": "string"
          },
          "risk_amount": {
            "title": "Risk Amount",
            "type": "string"
          }
        },
        "required": [
          "qty",
          "notional",
          "risk_amount",
          "pct_of_equity"
        ],
        "title": "PositionSizeOut",
        "type": "object"
      },
      "PositionSizeRequest": {
        "additionalProperties": false,
        "description": "Inputs for a position-sizing computation. All money/price inputs are decimals\nsent as JSON numbers; the server returns Decimal STRINGS (invariant #2) so the\nclient never does money math.\n\n``stop`` is the protective-stop PRICE (used to derive risk-per-share for the\n``risk_per_trade`` method); ``atr`` is the absolute ATR value for ``atr`` sizing.",
        "properties": {
          "atr": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atr"
          },
          "entry": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Entry"
          },
          "equity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Equity"
          },
          "method": {
            "enum": [
              "fixed_units",
              "percent_equity",
              "risk_per_trade",
              "atr"
            ],
            "title": "Method",
            "type": "string"
          },
          "stop": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop"
          },
          "value": {
            "exclusiveMinimum": 0.0,
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "method",
          "value",
          "equity",
          "entry"
        ],
        "title": "PositionSizeRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Trading Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/calc/position-size": {
      "post": {
        "description": "Position-sizing GUIDANCE from the engine's own ``_size()`` math. This is\nNOT an executable order \u2014 live trading is gated (POST /live/orders \u2192 403).",
        "operationId": "position_size_calc_position_size_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionSizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionSizeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Position Size",
        "tags": [
          "calc"
        ]
      }
    },
...
    "/paper/deploy": {
      "post": {
        "operationId": "deploy_paper_deploy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deploy",
        "tags": [
          "paper-trading"
        ]
      }
    },
...
    "/paper/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_paper_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "paper-trading"
        ]
      }
    },
    "/paper/sessions/{session_id}/run": {
      "post": {
        "operationId": "run_now_paper_sessions__session_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Now",
        "tags": [
          "paper-trading"
        ]
      }
    },
    "/paper/sessions/{session_id}/stop": {
      "post": {
        "operationId": "stop_paper_sessions__session_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop",
        "tags": [
          "paper-trading"
        ]
      }
    },
...
  }
}